1. bugfix, use MAC of gateway while replying in the background 2. save relay configuration
git-svn-id: http://svn.code.sf.net/p/vpcs/code/trunk@99 285c826d-31ef-41a9-a5b6-d2a4949dcdf2

+++ src/vpcs.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2007-2013, Paul Meng (mirnshi@gmail.com)
+ * Copyright (c) 2007-2014, Paul Meng (mirnshi@gmail.com)
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without 
@@ -80,7 +80,7 @@ typedef struct {
 	u_int ip;		/* local host ip */
 	int cidr;		/* local host ip netmask */
 	u_char mac[6];		/* local host mac */
-	u_long gw;		/* default gateway ip */
+	u_int gw;		/* default gateway ip */
 	u_char gmac[6];		/* destination host mac */
 	dhcp dhcp;				
 	u_int lease;		/* dhcp lease time */
